fixed ambigous operator error

+++ src/main.cpp
@@ -207,7 +207,7 @@ bool coord_out_of_range(settings_t& s, mc::utils::level_coord& coord)
 
 template<typename T>
 void cout_dot(T total) {
-  if (total == 0) out << " done!";
+  if ((unsigned int) total == 0) out << " done!";
   else out << "." << flush;
 }
